Home

PostgresCatalog.jl

Overview

PostgresCatalog is a Julia library for introspecting Postgres databases and generating models of the database structure.

Installation

Use the Julia package manager.

julia> using Pkg
julia> Pkg.add("PostgresCatalog")

Usage Guide

We will demonstrate how to use PostgresCatalog on a sample database schema containing just one table.

using LibPQ

conn = LibPQ.Connection("")

execute(conn, "BEGIN")

execute(conn,
        """
        CREATE TYPE patient_sex_enum AS ENUM ('male', 'female', 'other', 'unknown');

        CREATE TABLE patient (
            id int4 NOT NULL,
            mrn text NOT NULL,
            sex patient_sex_enum NOT NULL DEFAULT 'unknown',
            mother_id int4,
            father_id int4,
            CONSTRAINT patient_uk UNIQUE (id),
            CONSTRAINT patient_pk PRIMARY KEY (mrn),
            CONSTRAINT patient_mother_fk FOREIGN KEY (mother_id) REFERENCES patient (id),
            CONSTRAINT patient_father_fk FOREIGN KEY (father_id) REFERENCES patient (id)
        );
        """)

To generate a model of the database structure, we use function PostgresCatalog.introspect, which returns a PostgresCatalog.PGCatalog object.

using PostgresCatalog

cat = PostgresCatalog.introspect(conn)
#-> DATABASE " … "

By traversing the catalog, we can obtain the models of database tables, which are represented as PostgresCatalog.PGTable objects.

scm = cat["public"]
#-> SCHEMA "public"

tbl = scm["patient"]
#-> TABLE "patient"

The table model contains information about its columns in the form of PostgresCatalog.PGColumn objects.

foreach(println, tbl)
#=>
COLUMN "patient"."id" "int4" NOT NULL
COLUMN "patient"."mrn" "text" NOT NULL
COLUMN "patient"."sex" "patient_sex_enum" NOT NULL
COLUMN "patient"."mother_id" "int4" NULL
COLUMN "patient"."father_id" "int4" NULL
=#

Properties of a table column are available as attributes on the corresponding model object.

col = tbl["sex"]
#-> COLUMN "patient"."sex" "patient_sex_enum" NOT NULL

col.name
#-> "sex"

col.type_
#-> TYPE "patient_sex_enum"

col.not_null
#-> true

col.default
#-> "'unknown'::patient_sex_enum"

Description of unique and foreign key constraints defined on the table is also available in the form of PostgresCatalog.PGUniqueKey and PostgresCatalog.PGForeignKey objects.

tbl.primary_key
#-> CONSTRAINT "patient"."patient_pk" PRIMARY KEY ("mrn")

foreach(println, tbl.unique_keys)
#=>
CONSTRAINT "patient"."patient_pk" PRIMARY KEY ("mrn")
CONSTRAINT "patient"."patient_uk" UNIQUE ("id")
=#

foreach(println, tbl.foreign_keys)
#=>
CONSTRAINT "patient"."patient_father_fk" FOREIGN KEY ("father_id") REFERENCES "patient" ("id")
CONSTRAINT "patient"."patient_mother_fk" FOREIGN KEY ("mother_id") REFERENCES "patient" ("id")
=#

API Reference

PostgresCatalog.introspect — Function.

introspect(conn::LibPQ.Connection) :: PGCatalog

Introspects a Postgres database and generates a model of the database structure.

source

PostgresCatalog.PGCatalog — Type.

Model of a Postgres database.

name: name of the database;

schemas: collection of schemas owned by the database.

source

PostgresCatalog.PGSchema — Type.

Model of a database schema.

catalog: database that owns the schema;

name: name of the schema;

comment: comment on the schema;

type: collection of types owned by the schema;

tables: collection of tables owned by the schema.

source

PostgresCatalog.PGType — Type.

Model of a type.

schema: schema that owns the type;

name: name of the type;

labels: vector of labels for an ENUM type; nothing otherwise;

comment: comment on the type;

columns: set of columns of this type.

source

PostgresCatalog.PGTable — Type.

Model of a table.

schema: schema that owns the table;

name: name of the table;

comment: comment on the table;

columns: collection of table columns;

primary_key: primary key of the table, if any;

unique_keys: collection of unique keys defined on the table;

foreign_keys: collection of foreign keys defined on the table;

referring_foreign_keys: set of foreign keys that refer to this table.

source

PostgresCatalog.PGColumn — Type.

Model of a column.

table: table that owns the column;

name: name of the column;

type_: type of the column;

not_null: set if the column has NOT NULL constraint;

default: SQL expression that calculates the default column value; or $nothing$;

unique_keys: set of unique keys that include this columns;

foreign_keys: set of foreign keys that include this column;

referring_foreign_keys: set of foreign keys that target this column.

source

PostgresCatalog.PGUniqueKey — Type.

Model of a unique key constraint.

table: table that owns the key;

name: name of the constraint;

columns: columns included to the key;

primary: set if this is the primary key;

comment: comment on the constraint.

source

PostgresCatalog.PGForeignKey — Type.

Model of a foreign key constraint.

table: table that owns the key;

name: name of the constraint;

columns: columns included to the key;

target_table: table targeted by the key;

target_columns: columns targeted by the key;

on_delete: ON DELETE action;

on_update: ON UPDATE action;

comment: comment on the constraint.

source